+++ NackUn/StudyFirst/app/src/main/java/sample/nackun/com/studyfirst/data/DataSource.kt
@@ -1,0 +1,6 @@
+package sample.nackun.com.studyfirst.data
+
+interface DataSource {
+    fun requestMarkets(marketLike: String)
+    fun requestTickers(query: String)
+}
