+++ NackUn/StudyFirst/app/src/main/java/sample/nackun/com/studyfirst/data/Repository.kt
@@ -1,0 +1,17 @@
+package sample.nackun.com.studyfirst.data
+
+import sample.nackun.com.studyfirst.data.remote.RemoteDataSource
+
+class Repository(
+    val remoteDataSource: DataSource
+) : DataSource
+{
+    override fun requestMarkets(marketLike: String) {
+        remoteDataSource.requestMarkets(marketLike)
+    }
+
+    override fun requestTickers(query: String) {
+        remoteDataSource.requestTickers(query)
+    }
+
+}
